Error stack safety

+++ arbitrator/arbutil/src/evm/mod.rs
@@ -80,7 +80,7 @@ pub struct EvmData {
 }
 
 /// Returns the minimum number of EVM words needed to store `bytes` bytes.
-pub fn evm_words(bytes: u64) -> u64 {
+pub fn evm_words(bytes: u32) -> u32 {
     match bytes % 32 {
         0 => bytes / 32,
         _ => bytes / 32 + 1,
